Separated event narration from command execution
