+++ src/App.tsx
@@ -9,7 +9,7 @@ import { StoryView } from './view/component/story/story-view';
 import { GAELIC_HELP_COMMAND } from './command/help-command';
 import { CommandInput } from './view/component/command-input';
 import { newGame } from './generation/game-generator';
-import { narrateRoom } from './narrator/gaelic-english-narrator';
+import { GAELIC_ENGLISH_NARRATOR, narrateRoom } from './narrator/gaelic-english-narrator';
 
 export const GLOBAL_HELP_PROMPT = {l1: "Type 'help' for help.", l2: `Clò-sgrìobh '${GAELIC_HELP_COMMAND}' airson cuideachadh.`}
 
@@ -32,9 +32,25 @@ function App() {
 
   // When a command is ented in the Command Input component, execute it
   let onEnterCommand = function(commandInput: string) {
-      let newState = executeCommand(commandInput, gameState, storyState);
-      setGameState(newState.gameState);
-      setStoryState(newState.storyState);
+      // Execute the command and determine the new state
+      let stateTransition = executeCommand(commandInput, gameState, storyState);
+
+      // Narrate the change in state
+      let eventNarration: Story = GAELIC_ENGLISH_NARRATOR.narrateEvent(
+        stateTransition.event,
+        gameState,
+        stateTransition.gameStateAfter
+      );
+      // Combine the previous story, the player's input, and the new story
+      setStoryState({
+        story: [
+            ...storyState.story,
+            StoryElement.userInput({input: commandInput}),
+            ...eventNarration
+        ]
+    });
+
+    setGameState(stateTransition.gameStateAfter);
   }
 
   return (
