+++ src/view/component/story/user-input-view.tsx
